Feedback from review.
